[WIP] Update COCOMetric to use MMEval
